+++ yang/yang-model-util/src/main/java/org/opendaylight/yangtools/yang/model/util/type/RestrictedUint64Type.java
@@ -7,20 +7,19 @@
  */
 package org.opendaylight.yangtools.yang.model.util.type;
 
-import java.math.BigInteger;
 import java.util.Collection;
 import org.eclipse.jdt.annotation.Nullable;
+import org.opendaylight.yangtools.yang.common.Uint64;
 import org.opendaylight.yangtools.yang.model.api.SchemaPath;
 import org.opendaylight.yangtools.yang.model.api.UnknownSchemaNode;
 import org.opendaylight.yangtools.yang.model.api.type.RangeConstraint;
 import org.opendaylight.yangtools.yang.model.api.type.Uint64TypeDefinition;
 
-final class RestrictedUint64Type extends AbstractRangeRestrictedType<Uint64TypeDefinition, BigInteger>
+final class RestrictedUint64Type extends AbstractRangeRestrictedType<Uint64TypeDefinition, Uint64>
         implements Uint64TypeDefinition {
-
     RestrictedUint64Type(final Uint64TypeDefinition baseType, final SchemaPath path,
-        final Collection<UnknownSchemaNode> unknownSchemaNodes,
-        final @Nullable RangeConstraint<BigInteger> rangeConstraint) {
+            final Collection<UnknownSchemaNode> unknownSchemaNodes,
+            final @Nullable RangeConstraint<Uint64> rangeConstraint) {
         super(baseType, path, unknownSchemaNodes, rangeConstraint);
     }
